a lot of performance optimization

+++ src/main/java/de/c4vxl/pipeline/TextGenerationPipeline.java
@@ -45,8 +45,8 @@ public class TextGenerationPipeline extends Pipeline {
     public String forward(String prompt, int max_new_tokens) { return this.forward(prompt, max_new_tokens, null); }
     public String forward(String prompt, int max_new_tokens, TextGenerationModel.GenerationStream stream) { return this.forward(prompt, this.temperature, max_new_tokens, this.blockSize, stream); }
     public String forward(String prompt, double temperature, int max_new_tokens, int block_size, TextGenerationModel.GenerationStream stream) {
-        Tensor<Integer> tokenized = this.tokenizer.forward(prompt);
-        Tensor<Integer> output = model.generate(tokenized, temperature, max_new_tokens, block_size, this.tokenizer.eosTokenID(), stream);
+        Tensor<Float> tokenized = this.tokenizer.forward(prompt);
+        Tensor<Float> output = model.generate(tokenized, temperature, max_new_tokens, block_size, this.tokenizer.eosTokenID(), stream);
         return this.tokenizer.decode(output);
     }
 }
